refactor: modifying path to Cart components.
